+++ CHANGELOG.md
@@ -5,7 +5,9 @@ Mapbox welcomes participation and contributions from everyone.
 ## main
 
 ### Features ✨ and improvements 🏁
-* Turf has been reexported. Developers no longer need to import Turf in their code base. If you are using `Turf.Geometry` directly in your code, you will need to `import Turf`. [#622](https://github.com/mapbox/mapbox-maps-ios/pull/622)
+
+* It is no longer necessary to `import Turf`. [#622](https://github.com/mapbox/mapbox-maps-ios/pull/622)
+* `Geometry` now refers to `Turf.Geometry` instead of `MapboxCommon.Geometry`. [#622](https://github.com/mapbox/mapbox-maps-ios/pull/622)
 
 ## 10.0.0-rc.7 - August 25, 2021